remove sidekiq
