+++ app/mailer_previews/application_mailer_preview.rb
@@ -1,7 +1,23 @@
-class ApplicationMailerPreview < ActionMailer::Preview
+# Preview all emails at http://localhost:3000/rails/mailers/application_mailer
 
+class ApplicationMailerPreview < ActionMailer::Preview
   def lease_reminder
     order = Order.first
     ApplicationMailer.lease_reminder(order.id)
+  end
+
+  def new_inquiry
+    inquiry = Inquiry.first
+    ApplicationMailer.new_inquiry(inquiry.id)
+  end
+
+  def new_request
+    request = Request.first
+    ApplicationMailer.new_request(request.id)
+  end
+  
+  def charge_reminder
+    charge = Charge.first
+    ApplicationMailer.charge_reminder(charge.id)
   end
 end
