+++ app/views/application_mailer/charge_reminder.html.slim
@@ -1,4 +1,4 @@
 p 
   | You have a charge that is due soon. To see the detail, click 
-  span = link_to("here", [:edit, @user, @order])
+  span = link_to("here", [:backend, @user, @order])
   | .
