+++ config/application.rb
@@ -49,6 +49,7 @@ module Homey
     
     config.assets.version = "1.0"
     
-    config.active_job.queue_adapter = :sidekiq
+    # config.active_job.queue_adapter = :sidekiq
+    config.active_job.queue_adapter = :inline
   end
 end
