+++ config/secrets.yml
@@ -1,3 +1,13 @@
+development: &development
+  domain_name: "example.com"
+  secret_key_base: abc
+  aws_access_key_id: <%= ENV["AWS_ACCESS_KEY_ID"] %>
+  aws_secret_access_key: <%= ENV["AWS_SECRET_ACCESS_KEY"] %>
+  weather_api_id: <%= ENV["WEATHER_API_ID"] %>
+
+test:
+  <<: *development
+
 production: &production
   domain_name: <%= ENV["DOMAIN_NAME"] %>
   secret_key_base: <%= ENV["SECRET_KEY_BASE"] %>
