+++ spec/models/request_spec.rb
@@ -7,13 +7,8 @@ RSpec.describe Request, type: :model do
   
   let(:request) { create(:request) }
   it "should create a mailer after create" do
-    Sidekiq::Worker.jobs.clear
+    ActionMailer::Base.deliveries.clear
     request.reload
-    jobs = Sidekiq::Worker.jobs
-    expect(jobs.length).to eq(1)
-    job = jobs[ActiveJob::QueueAdapters::SidekiqAdapter::JobWrapper].first
-    expect(job["at"].to_f - job["enqueued_at"].to_f).to eq(60.0)
-    args = job["args"].first["arguments"]
-    expect(args[0..1] << args[3]).to eq(["ApplicationMailer", "new_request", request.id])
+    expect(ActionMailer::Base.deliveries.first.subject).to eq("Prospective Customer Left Request")
   end
 end
